+++ tests/test_unit/test_SWC_Blueprint.py
@@ -1,21 +1,9 @@
 from pathlib import Path
-import pytest
+
 from rsp_vision.objects.SWC_Blueprint import (
-    SWC_Blueprint_Spec,
-    SubjectFolder,
     SessionFolder,
+    SubjectFolder,
 )
-
-
-@pytest.fixture
-def blueprint_spec(tmp_path):
-    spec = SWC_Blueprint_Spec(
-        project_name="my_project",
-        raw_data=True,
-        derivatives=True,
-        local_path=tmp_path,
-    )
-    return spec
 
 
 def test_SWC_Blueprint_Spec(tmp_path, blueprint_spec):
@@ -30,7 +18,10 @@ def test_SubjectFolder(tmp_path, blueprint_spec, folder_naming_specs):
         Path(subject_folder.sub_folder_path).mkdir(parents=True, exist_ok=True)
 
         assert (
-            tmp_path / "my_project" / "derivatives" / subject_folder.sub_folder_name
+            tmp_path
+            / "my_project"
+            / "derivatives"
+            / subject_folder.sub_folder_name
         ).exists()
 
         assert (
@@ -46,27 +37,29 @@ def test_SessionFolder(tmp_path, blueprint_spec, folder_naming_specs):
         Path(session_folder.ses_folder_path).mkdir(parents=True, exist_ok=True)
 
         assert (
-            tmp_path / "my_project" / "derivatives" / subject_folder.sub_folder_name / session_folder.ses_folder_name
+            tmp_path
+            / "my_project"
+            / "derivatives"
+            / subject_folder.sub_folder_name
+            / session_folder.ses_folder_name
         ).exists()
 
         monitor = fns.monitor_position[8:].replace("_", "-").replace("-", "")
 
-        test_name = f"ses-000_hemisphere-{fns.hemisphere}_region-{fns.brain_region}_monitor-{monitor}"
+        test_name = (
+            f"ses-000_hemisphere-{fns.hemisphere}_"
+            + f"region-{fns.brain_region}_monitor-{monitor}"
+        )
 
         if (fns.fov is None) and (fns.cre is None):
-            assert (
-                session_folder.ses_folder_name
-                == test_name
-            )
+            assert session_folder.ses_folder_name == test_name
         elif (fns.fov is not None) and (fns.cre is None):
             assert (
-                session_folder.ses_folder_name
-                == test_name + f"_fov-{fns.fov}"
+                session_folder.ses_folder_name == test_name + f"_fov-{fns.fov}"
             )
         elif (fns.fov is None) and (fns.cre is not None):
             assert (
-                session_folder.ses_folder_name
-                == test_name + f"_cre-{fns.cre}"
+                session_folder.ses_folder_name == test_name + f"_cre-{fns.cre}"
             )
         else:
             assert (
